Remove global variables. Refactor UI code into separate ReactiveChat class.
