+++ src/UI/reactive_chat.py
@@ -1,0 +1,132 @@
+
+
+import param
+import panel as pn
+import asyncio
+import re
+from src.UI.avatar import avatar
+import src.Agents.agents as agents
+from src import globals as globals
+
+class ReactiveChat(param.Parameterized):
+    def __init__(self, groupchat_manager, **params):
+        super().__init__(**params)
+        
+        pn.extension(design="material")
+
+        self.groupchat_manager = groupchat_manager
+
+        # Learn tab
+        self.LEARN_TAB_NAME = "LearnTab"
+        self.learn_tab_interface = pn.chat.ChatInterface(callback=self.a_learn_tab_callback, name=self.LEARN_TAB_NAME)
+
+        # Dashboard tab
+        self.dashboard_view = pn.pane.Markdown(f"Total messages: {len(self.groupchat_manager.groupchat.messages)}")
+        
+        # Progress tab
+        self.progress_text = pn.pane.Markdown(f"**Student Progress**")
+        self.progress = 0
+        self.max_questions = 10
+        self.progress_bar = pn.widgets.Progress(name='Progress', value=self.progress, max=self.max_questions)        
+        self.progress_info = pn.pane.Markdown(f"{self.progress} out of {self.max_questions}", width=60)
+
+        # Model tab. Capabilities for the LearnerModel
+        self.MODEL_TAB_NAME = "ModelTab"
+        self.model_tab_interface = pn.chat.ChatInterface(callback=self.a_model_tab_callback, name=self.MODEL_TAB_NAME)
+        self.button_update_learner_model = pn.widgets.Button(name='Update Learner Model', button_type='primary')
+        self.button_update_learner_model.on_click(self.handle_button_update_model)
+        self.is_model_tab = False
+        #self.learner_model_interface.send("What is the student's current capabilities", user="System", respond=True)
+
+        # TODO: Consider whether groupchat_manager or this class should manage the chat_interface
+        #       Currently, I have placed it in CustomGroupChatManager
+        self.groupchat_manager.chat_interface = self.learn_tab_interface  # default chat tab
+
+    ############ tab1: Learn interface
+    async def a_learn_tab_callback(self, contents: str, user: str, instance: pn.chat.ChatInterface):
+        '''
+            All panel callbacks for the learn tab come through this callback function
+            Because there are two chat panels, we need to save the instance
+            Then, when update is called, check the instance name
+        '''                      
+        self.groupchat_manager.chat_interface = instance
+        if not globals.initiate_chat_task_created:
+            asyncio.create_task(self.groupchat_manager.delayed_initiate_chat(agents.tutor, self.groupchat_manager, contents))  
+        else:
+            if globals.input_future and not globals.input_future.done():                
+                globals.input_future.set_result(contents)                 
+            else:
+                print("No input being awaited.")
+    
+    def update_learn_tab(self, recipient, messages, sender, config):
+        if self.groupchat_manager.chat_interface.name is not self.LEARN_TAB_NAME: return
+        last_content = messages[-1]['content'] 
+        if all(key in messages[-1] for key in ['name']):
+            self.learn_tab_interface.send(last_content, user=messages[-1]['name'], avatar=avatar[messages[-1]['name']], respond=False)
+        else:
+            self.learn_tab_interface.send(last_content, user=recipient.name, avatar=avatar[recipient.name], respond=False)
+        
+    ########## tab2: Dashboard
+    def update_dashboard(self):
+        self.dashboard_view.object = f"Total messages: {len(self.groupchat_manager.groupchat.get_messages())}"
+
+    ########### tab3: Progress
+    def update_progress(self, contents, user):
+        # Parse the agent's output for keywords                 
+        if user == "LevelAdapterAgent":            
+            pattern = re.compile(r'\b(correct|correctly|verified|yes|well done|excellent|successfully|that\'s right|good job|excellent|right|good|affirmative)\b', re.IGNORECASE)            
+            is_correct = pattern.search(contents)
+            if is_correct:
+               print("################ CORRECT ANSWER #################")
+               if self.progress < self.max_questions:  
+                    self.progress += 1
+                    self.progress_bar.value = self.progress
+                    self.progress_info.object = f"**{self.progress} out of {self.max_questions}**"
+
+            else:
+                print("################ WRONG ANSWER #################")
+
+    ########## tab4: Learner Model
+    async def handle_button_update_model(self, event=None):
+        self.groupchat_manager.chat_interface = self.model_tab_interface
+        await self.a_update_model()
+     
+    async def a_update_model(self):
+        '''
+            This is a long latency operation therefore async
+        '''
+        if self.groupchat_manager.chat_interface.name is not self.MODEL_TAB_NAME: return
+        messages = self.groupchat_manager.groupchat.get_messages()
+        for m in messages:
+            agents.learner_model.send(m, recipient=agents.learner_model, request_reply=False)
+        await agents.learner_model.a_send("What is the student's current capabilities", recipient=agents.learner_model, request_reply=True)
+        response = agents.learner_model.last_message(agent=agents.learner_model)["content"]
+        self.model_tab_interface.send(response, user=agents.learner_model.name,avatar=avatar[agents.learner_model.name])
+
+
+    async def a_model_tab_callback(self, contents: str, user: str, instance: pn.chat.ChatInterface):
+        '''
+            Receive any input from the ChatInterface of the Model tab
+        '''
+        self.groupchat_manager.chat_interface = instance
+        if user == "System" or user == "User":
+            response = agents.learner_model.last_message(agent=agents.learner_model)["content"]
+            self.learn_tab_interface.send(response, user=agents.learner_model.name,avatar=avatar[agents.learner_model.name])
+    
+    ########## Create the "windows" and draw the tabs
+    def draw_view(self):         
+        tabs = pn.Tabs(  
+            ("Learn", pn.Column(self.learn_tab_interface)),
+            ("Dashboard", pn.Column(self.dashboard_view)),
+            ("Progress", pn.Column(
+                    self.progress_text,
+                    pn.Row(                        
+                        self.progress_bar,
+                        self.progress_info
+                    ))),
+            ("Model", pn.Column(
+                      pn.Row(self.button_update_learner_model),
+                      pn.Row(self.model_tab_interface)))            
+        )
+        return tabs
+
